backend/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
pandas==2.1.4
reportlab==4.0.8
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
dj-database-url==2.1.0
